fix: changed implementation

+++ src/internal/models/types.go
@@ -6,10 +6,10 @@ import (
 )
 
 type product struct {
-	Name string `json:"name" bson:"name"`
-	Price uint64 `json:"price" bson:"price"`
+	Name  string `json:"name" bson:"name"`
+	Price uint32 `json:"price" bson:"price"`
 }
 
-func GenerateProduct() product{
-	return product{Name: strconv.Itoa(rand.Int()),Price: rand.Uint64()}
+func GenerateProduct() product {
+	return product{Name: strconv.Itoa(rand.Int()), Price: rand.Uint32()}
 }
